requirements.txt
# Core document processing
PyPDF2==3.0.1
python-docx==0.8.11

# Google Gemini API for embeddings
google-generativeai==0.3.2

# Configuration management
python-dotenv==1.0.0

# PostgreSQL database
psycopg2-binary==2.9.7
